+++ src/traces/bar/defaults.js
@@ -7,7 +7,7 @@ var Registry = require('../../registry');
 var handleXYDefaults = require('../scatter/xy_defaults');
 var handlePeriodDefaults = require('../scatter/period_defaults');
 var handleStyleDefaults = require('./style_defaults');
-var getAxisGroup = require('../../plots/cartesian/constraints').getAxisGroup;
+var handleGroupingDefaults = require('./grouping_defaults');
 var attributes = require('./attributes');
 
 var coerceFont = Lib.coerceFont;
@@ -56,45 +56,6 @@ function supplyDefaults(traceIn, traceOut, defaultColor, layout) {
     errorBarsSupplyDefaults(traceIn, traceOut, lineColor || Color.defaultLine, {axis: 'x', inherit: 'y'});
 
     Lib.coerceSelectionMarkerOpacity(traceOut, coerce);
-}
-
-function handleGroupingDefaults(traceIn, traceOut, fullLayout, coerce) {
-    var orientation = traceOut.orientation;
-    // N.B. grouping is done across all trace types that support it
-    var posAxId = traceOut[{v: 'x', h: 'y'}[orientation] + 'axis'];
-    var groupId = getAxisGroup(fullLayout, posAxId) + orientation;
-
-    var alignmentOpts = fullLayout._alignmentOpts || {};
-    var alignmentgroup = coerce('alignmentgroup');
-
-    var alignmentGroups = alignmentOpts[groupId];
-    if(!alignmentGroups) alignmentGroups = alignmentOpts[groupId] = {};
-
-    var alignmentGroupOpts = alignmentGroups[alignmentgroup];
-
-    if(alignmentGroupOpts) {
-        alignmentGroupOpts.traces.push(traceOut);
-    } else {
-        alignmentGroupOpts = alignmentGroups[alignmentgroup] = {
-            traces: [traceOut],
-            alignmentIndex: Object.keys(alignmentGroups).length,
-            offsetGroups: {}
-        };
-    }
-
-    var offsetgroup = coerce('offsetgroup');
-    var offsetGroups = alignmentGroupOpts.offsetGroups;
-    var offsetGroupOpts = offsetGroups[offsetgroup];
-
-    if(offsetgroup) {
-        if(!offsetGroupOpts) {
-            offsetGroupOpts = offsetGroups[offsetgroup] = {
-                offsetIndex: Object.keys(offsetGroups).length
-            };
-        }
-
-        traceOut._offsetIndex = offsetGroupOpts.offsetIndex;
-    }
 }
 
 function crossTraceDefaults(fullData, fullLayout) {
@@ -172,6 +133,5 @@ function handleText(traceIn, traceOut, layout, coerce, textposition, opts) {
 module.exports = {
     supplyDefaults: supplyDefaults,
     crossTraceDefaults: crossTraceDefaults,
-    handleGroupingDefaults: handleGroupingDefaults,
     handleText: handleText
 };
